+++ eidas_node/proxy_service/settings.py
@@ -30,14 +30,16 @@ class ProxyServiceSettings(AppSettings):
             settings=dict(
                 # required=True leads to a strange error:
                 # "REQUEST_SIGNATURE setting is missing required item 'REQUEST_SIGNATURE'"
-                key_file=StringSetting(min_length=1),
+                key_source=StringSetting(min_length=1),
+                key_location=StringSetting(min_length=1),
                 cert_file=StringSetting(min_length=1),
                 signature_method=StringSetting(default='RSA_SHA512', min_length=1),
                 digest_method=StringSetting(default='SHA512', min_length=1),
             ),
             # https://github.com/pawamoy/django-appsettings/issues/91
             required=True),
-        key_file=StringSetting(),
+        key_source=StringSetting(),
+        key_location=StringSetting(),
         cert_file=StringSetting(),
     ), required=True)
     light_storage = NestedDictSetting(settings=dict(
@@ -70,8 +72,10 @@ def check_settings():
     ProxyServiceSettings.check()
     signature = PROXY_SERVICE_SETTINGS.identity_provider['request_signature']
     # If one of the files is set, the other must be set as well
-    if bool(signature.get('key_file')) != bool(signature.get('cert_file')):
-        raise ImproperlyConfigured('Both PROXY_SERVICE_IDENTITY_PROVIDER.REQUEST_SIGNATURE.KEY_FILE and '
+    if not (bool(signature.get('key_source')) == bool(signature.get('key_location'))
+            == bool(signature.get('cert_file'))):
+        raise ImproperlyConfigured('PROXY_SERVICE_IDENTITY_PROVIDER.REQUEST_SIGNATURE.KEY_SOURCE, '
+                                   'PROXY_SERVICE_IDENTITY_PROVIDER.REQUEST_SIGNATURE.KEY_LOCATION and '
                                    'PROXY_SERVICE_IDENTITY_PROVIDER.REQUEST_SIGNATURE.CERT_FILE must be set.')
 
     auxiliary_required = PROXY_SERVICE_SETTINGS.transient_name_id_fallback or PROXY_SERVICE_SETTINGS.track_country_code