+++ eidas_node/tests/proxy_service/test_settings.py
@@ -5,13 +5,14 @@ from django.core.exceptions import ImproperlyConfigured
 from django.test import SimpleTestCase, override_settings
 
 from eidas_node.proxy_service.settings import check_settings
-from eidas_node.tests.constants import CERT_FILE, KEY_FILE
+from eidas_node.tests.constants import CERT_FILE, KEY_LOCATION, KEY_SOURCE
 
 PROXY_SERVICE_IDENTITY_PROVIDER: Dict[str, Any] = {
     'ENDPOINT': 'https://test.example.net/identity-provider-endpoint',
     'REQUEST_ISSUER': 'https://test.example.net/saml/idp.xml',
     'REQUEST_SIGNATURE': {
-        'KEY_FILE': KEY_FILE,
+        'KEY_SOURCE': KEY_SOURCE,
+        'KEY_LOCATION': KEY_LOCATION,
         'CERT_FILE': CERT_FILE,
         'SIGNATURE_METHOD': 'RSA_SHA1',
         'DIGEST_METHOD': 'SHA1',
@@ -33,13 +34,14 @@ class TestCheckSettings(SimpleTestCase):
     def test_check_settings_signature_no_cert_and_no_key(self):
         identity_provider = deepcopy(PROXY_SERVICE_IDENTITY_PROVIDER)
         del identity_provider['REQUEST_SIGNATURE']['CERT_FILE']
-        del identity_provider['REQUEST_SIGNATURE']['KEY_FILE']
+        del identity_provider['REQUEST_SIGNATURE']['KEY_SOURCE']
+        del identity_provider['REQUEST_SIGNATURE']['KEY_LOCATION']
         with override_settings(PROXY_SERVICE_IDENTITY_PROVIDER=identity_provider):
             check_settings()
 
     def test_check_settings_signature_cert_and_no_key(self):
         identity_provider = deepcopy(PROXY_SERVICE_IDENTITY_PROVIDER)
-        del identity_provider['REQUEST_SIGNATURE']['KEY_FILE']
+        del identity_provider['REQUEST_SIGNATURE']['KEY_LOCATION']
         with override_settings(PROXY_SERVICE_IDENTITY_PROVIDER=identity_provider):
             self.assertRaises(ImproperlyConfigured, check_settings)